updata ticket list api
