+++ config/app.js
@@ -27,6 +27,7 @@ import activityRouters from "../routes/activity.js";
 import venueRouters from "../routes/venue.js";
 import settingRouters from "../routes/activitySetting.js";
 import ticketTypeRouters from "../routes/ticketType.js"
+import ticketListRouters from "../routes/ticketList.js"
 import partnerRouters from "../routes/partner.js";
 import faqRouters  from "../routes/faq.js";
 
@@ -115,6 +116,7 @@ app.use("/api/v1/activities", activityRouters);
 app.use("/api/v1/venues", venueRouters);
 app.use("/api/v1/settings", settingRouters);
 app.use("/api/v1/ticketTypes",ticketTypeRouters);
+app.use("/api/v1/ticketList",ticketListRouters);
 app.use("/api/v1/partners", partnerRouters);
 app.use("/api/v1/faqs", faqRouters);
 app.use("/api/v1", otherRouters);
